Yellow Magic Orchestra Website
This is the first thing I ever programmed: an informational website on the Japanese 80s band, Yellow Magic Orchestra. It was made as part of the "Practice Create Task" for my highschool's AP CS Principals class.

It uses HTML, CSS, and some Javascript which doesn't work.

This poorly made website was my introduction to programming.